requirements.txt
asgiref==3.5.2
autopep8==2.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==39.0.0
defusedxml==0.7.1
dj-rest-auth==2.2.7
Django==4.1.4
django-admin-generator==2.6.0
django-allauth==0.52.0
django-baton==2.6.0
django-rest-registration==0.7.3
django-utils2==2.12.1
djangorestframework==3.14.0
idna==3.4
oauthlib==3.2.2
Pillow==9.3.0
pycodestyle==2.10.0
pycparser==2.21
PyJWT==2.6.0
python-utils==3.5.2
python3-openid==3.2.0
pytz==2022.7
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.7
urllib3==1.26.14
